[evrironment] 1. add driver interface header file, 2. add autotradingsystem.cpp
