+++ DAY10_stocktrading/autotradingsystem.cpp
@@ -1,0 +1,2 @@
+#pragma once
+#include "autotradingsystem.h"
