+++ DAY10_stocktrading/autotradingsystem.h
@@ -1,1 +1,10 @@
 #pragma once
+#include "StockBrokerDriverInterface.h"
+
+class AutoTradingSystem
+{
+public:
+	
+private:
+	StockBrokerDriverInterface* broker;
+};
