+++ DAY10_stocktrading/main.cpp
@@ -7,15 +7,6 @@
 using namespace testing;
 using namespace std;
 
-class StockBrokerDriverInterface {
-public:
-    virtual std::string selectStockBrocker() const = 0;
-    virtual bool login(const std::string& id, const std::string& pass) = 0;
-    virtual bool buy(const std::string& code, int price, int quantity) = 0;
-    virtual bool sell(const std::string& code, int price, int quantity) = 0;
-    virtual int getPrice(const std::string& code) = 0;
-};
-
 // Mock 클래스 정의
 class MockStockBrokerDriver : public StockBrokerDriverInterface {
 public:
